Skip delay of 0, deliver actionmailer
Turns out the Sidekiq Delayed extensions were hiding some complexity related to figuring out what kind of class the module was mixed into. It would appropriately use `DelayedModel` , `DelayedMailer` etc depending on what type of object it was wrapping.

What we should do here is step back and re-architect how the static subscribers are created and handled. There should be a few different classes depending on the options passed in to `on_event` to minimize the method-level branches. This patch works for now, but clearly we don't have tests that will cover that kind of architecture change.

Will consider it.